Inventory #2, Cells, UI, Cameras
Worked on Inventory and a lot of small things:

Inventory: added a lot of functionality for the Inventory UI - UI nearly completed

Cells: Added Water cells to the game. Added rule Tile for sand cells (incomplete). Aded Tilemap for Water.
-> Problems with cells in different Tilemaps

UI: changed UI in the 3 scenes to scale with screensize and adapted the scales

Cameras: implemented individual cameras for each player in game scene